requirements.txt
django>=5.2,<6.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
google-genai
